Layout and performance updates
Due to performance issues (using 5+Cores for a single process) there was need for a performance update. 
Throughout this performance update the documentation/layout of the code has also been updated to be a bit neater, consistent, and easier to understand. 

The major approach to conducting the simulation and it's rates has not changed.